+++ Configuration/TCA/Overrides/tt_content.php
@@ -13,6 +13,30 @@ use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 defined('TYPO3') || die();
 
 (static function (string $contentType = 'tx_jobrouterdata_table', $icon = 'jobrouter-data-ce-table'): void {
+    $tableColumn = [
+        'tx_jobrouterdata_table' => [
+            'label' => Extension::LANGUAGE_PATH_CONTENT_ELEMENT . ':table',
+            'config' => [
+                'type' => 'select',
+                'renderType' => 'selectSingle',
+                'items' => [
+                    [
+                        'label' => '',
+                        'value' => '',
+                    ],
+                ],
+                'foreign_table' => 'tx_jobrouterdata_domain_model_table',
+                'foreign_table_where' => 'AND {#tx_jobrouterdata_domain_model_table}.{#type}=1 ORDER BY name',
+                'required' => true,
+            ],
+        ],
+    ];
+
+    ExtensionManagementUtility::addTCAcolumns(
+        'tt_content',
+        $tableColumn,
+    );
+
     ExtensionManagementUtility::addPlugin(
         [
             'label' => Extension::LANGUAGE_PATH_CONTENT_ELEMENT . ':ce.title',
@@ -25,24 +49,13 @@ defined('TYPO3') || die();
         Extension::KEY,
     );
 
-    ExtensionManagementUtility::addPiFlexFormValue(
-        '*',
-        'FILE:EXT:' . Extension::KEY . '/Configuration/FlexForms/Table.xml',
-        $contentType,
-    );
-
     $GLOBALS['TCA']['tt_content']['types'][$contentType] = [
-        'columnsOverrides' => [
-            'pi_flexform' => [
-                'label' => Extension::LANGUAGE_PATH_CONTENT_ELEMENT . ':table',
-            ],
-        ],
         'showitem' => '
             --div--;LLL:EXT:core/Resources/Private/Language/Form/locallang_tabs.xlf:general,
                 --palette--;;general,
                 --palette--;;headers,
             --div--;' . Extension::LANGUAGE_PATH_CONTENT_ELEMENT . ':table,
-                pi_flexform,
+                tx_jobrouterdata_table,
             --div--;LLL:EXT:frontend/Resources/Private/Language/locallang_ttc.xlf:tabs.appearance,
                 --palette--;;frames,
                 --palette--;;appearanceLinks,