+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Compatibility with TYPO3 v14
 
+### Changed
+- The table field of the content element was moved from FlexForms to a native database field
+
 ### Removed
 - Compatibility with TYPO3 v12 (#26)